requirements.txt
fastapi
pydantic
uvicorn
httpx
pytest==8.3.5
openai

python-dotenv
